refactor(mobile-push): isolate background worker
