+++ example.env
@@ -15,19 +15,9 @@ REDDIT_CLIENT_ID=
 REDDIT_CLIENT_SECRET=
 REDDIT_USERNAME=
 REDDIT_PASSWORD=
-# mobile_push domain (admin Posts API + scheduler). ENCRYPTION_KEY (shared
-# with clashking-api) is used to decrypt mobile_push_devices tokens unless
-# MOBILE_PUSH_TOKEN_KEY overrides it.
+# mobile_push background worker. ENCRYPTION_KEY (shared with clashking-api)
+# decrypts mobile_push_devices tokens unless MOBILE_PUSH_TOKEN_KEY overrides it.
 ENCRYPTION_KEY=
-# Discord OAuth is the only admin authentication method. The callback must be
-# registered exactly in the Discord Developer Portal.
-MOBILE_PUSH_DISCORD_CLIENT_ID=
-MOBILE_PUSH_DISCORD_CLIENT_SECRET=
-MOBILE_PUSH_DISCORD_REDIRECT_URL=http://localhost:8090/auth/discord/callback
-MOBILE_PUSH_ADMIN_PANEL_URL=http://localhost:5173
-# Internal operational metrics endpoint shown in the admin panel.
-MOBILE_PUSH_PROXY_STATS_URL=https://proxy.clashk.ing/stats
-BUNNY_ACCESS_KEY=
 MOBILE_PUSH_APNS_BEARER_TOKEN=
 MOBILE_PUSH_APNS_BUNDLE_ID=
 # Preferred: configure Application Default Credentials. Locally, run
