+++ scripts/mobile_push.go
@@ -1,11 +1,8 @@
 package scripts
 
 import (
-	"bytes"
 	"context"
 	"errors"
-	"fmt"
-	"net/http"
 	"time"
 
 	"clashking_tracking/internal/platform"
@@ -14,10 +11,8 @@ import (
 
 const mobilePushDomainName = "mobile_push"
 
-// mobilePushDomain serves the admin Posts API and, on the same interval
-// loop shape as giveaways.go/scheduled.go, publishes scheduled posts whose
-// starts_at has passed and expires live posts whose ends_at has passed —
-// with no admin action required for either.
+// mobilePushDomain publishes scheduled posts, retries push deliveries, runs
+// recurring campaigns, and expires live posts without exposing an HTTP API.
 type mobilePushDomain struct {
 	store mobilePushStore
 }
@@ -36,19 +31,6 @@ func (d *mobilePushDomain) Run(ctx context.Context, app *platform.App) error {
 	}
 	defer store.Close()
 	d.store = store
-
-	server := newMobilePushHTTPServer(app, store)
-	go func() {
-		if err := server.ListenAndServe(); err != nil && !errors.Is(err, http.ErrServerClosed) {
-			app.Logger.Error("mobile_push: http server exited", "err", err)
-		}
-	}()
-	defer func() {
-		shutdownCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-		_ = server.Shutdown(shutdownCtx)
-	}()
-	app.Logger.Info("mobile_push: http server starting", "addr", app.Config.MobilePushHTTPAddr)
 
 	interval := time.Duration(app.Config.MobilePushScanSeconds) * time.Second
 	for {
@@ -73,9 +55,6 @@ func (d *mobilePushDomain) Run(ctx context.Context, app *platform.App) error {
 func validateMobilePushConfig(cfg platform.Config) error {
 	if cfg.MobilePushScanSeconds <= 0 {
 		return errors.New("mobile_push.scan_seconds must be greater than zero")
-	}
-	if cfg.MobilePushHTTPAddr == "" {
-		return errors.New("mobile_push.http_addr is required")
 	}
 	if !cfg.DryRun && !cfg.MockDB && cfg.TimescaleURL == "" {
 		return errors.New("TIMESCALE_URL is required for mobile_push")
@@ -166,9 +145,7 @@ type publishResult struct {
 	PushSkipped int `json:"push_skipped"`
 }
 
-// publishAndNotify is the single publish-and-maybe-push path shared by the
-// scheduler's due-post scan and the "Publish now" HTTP handler, so both
-// behave identically.
+// publishAndNotify is the worker's single publish-and-maybe-push path.
 func publishAndNotify(ctx context.Context, app *platform.App, store mobilePushStore, post models.AdminPost) (publishResult, error) {
 	updated, err := store.MarkPublished(ctx, post.ID)
 	if err != nil {
@@ -243,30 +220,4 @@ func sendLocalizedPush(ctx context.Context, app *platform.App, devices []models.
 		skipped += groupSkipped
 	}
 	return sent, skipped
-}
-
-// bunnyUpload mirrors clashking-api/internal/routes/cdn.go's BunnyCDN PUT
-// technique exactly, targeting the same storage zone and public CDN host
-// under an admin-posts/ prefix, so it needs no new hosting infrastructure —
-// just its own access key.
-func bunnyUpload(accessKey, path, ext string, data []byte) (string, error) {
-	fullPath := fmt.Sprintf("%s.%s", path, ext)
-	uploadURL := fmt.Sprintf("https://storage.bunnycdn.com/clashking-files/%s", fullPath)
-
-	req, err := http.NewRequest(http.MethodPut, uploadURL, bytes.NewReader(data))
-	if err != nil {
-		return "", err
-	}
-	req.Header.Set("AccessKey", accessKey)
-	req.Header.Set("Content-Type", "application/octet-stream")
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode >= 300 {
-		return "", fmt.Errorf("bunnycdn upload failed: status %d", resp.StatusCode)
-	}
-	return fmt.Sprintf("https://cdn.clashk.ing/%s", fullPath), nil
 }
